refactor(orbit-controls): implement threejs autoRotate
